fix(symmetry): write vld artifacts in double, not lossy single JSON
Fixing the reader-side class mismatch (bf791a3) exposed a real defect
underneath it. With both sides now double, the comparison failed on
precision instead:

    Index          Actual          Expected         Error         AbsTol
      7      -0.303353995084763   -0.303354    4.9152e-09         1e-09
     61      -0.286873996257782   -0.286874    3.7422e-09         1e-09

vld_example.vlh declares precision int16, so ndr.reader.vld returns
SINGLE, and jsonencode formats a single with single-precision width. The
artifact on disk kept only ~6 significant digits of a value the reader
had produced to full single precision.

That matters beyond this test: readData.json is the cross-language
contract, so NDR-python was going to be checked against a 6-digit
rounding rather than against what the MATLAB reader actually returns. The
tolerance was not too tight -- the artifact was lossy.

Cast to double before encoding, which is exact for single input and makes
jsonencode emit full precision. Do the same in the intan generator: it
returns double today so the cast is a no-op, but it closes the same trap
rather than leaving it armed in the one file that happened not to trip.
readData_tiffstack.m and readData_prairieview.m already cast this way.

Note the tolerance is deliberately left at 1e-9. Loosening it to 1e-6
would have turned this green while keeping a lossy artifact, which is the
opposite of what the symmetry suite is for.

+++ tools/tests/+ndr/+symmetry/+makeArtifacts/+reader/readData.m
@@ -67,7 +67,14 @@ classdef readData < matlab.unittest.TestCase
                 epochstreams, epoch_select, s0, s1);
 
             readStruct = struct();
-            readStruct.ai_channel_1_samples_1_100 = data(:)';
+            % Cast to double before encoding for the same reason as the vld
+            % generator: jsonencode formats a single with single-precision
+            % width and silently truncates the artifact to ~6 significant
+            % digits. The intan reader returns double today, so this is a
+            % no-op now, but it stops the trap from reappearing if the
+            % reader's return type ever changes. (readData_tiffstack.m and
+            % readData_prairieview.m already do this.)
+            readStruct.ai_channel_1_samples_1_100 = double(data(:)');
 
             readJson = jsonencode(readStruct, 'ConvertInfAndNaN', true, 'PrettyPrint', true);
             fid = fopen(fullfile(artifactDir, 'readData.json'), 'w');
